fix: 🐛 ID 타입 Integer에서 Long으로 수정

+++ back/myini-api/src/main/java/com/ssafy/myini/security/oauth/CustomOAuth2User.java
@@ -16,7 +16,7 @@ import java.util.Map;
 @Data
 public class CustomOAuth2User implements OAuth2User {
 
-    private int userId;
+    private Long userId;
     private Provider userProvider;
     private String userProviderId;
     private String userName;
@@ -26,7 +26,7 @@ public class CustomOAuth2User implements OAuth2User {
     private Collection<? extends GrantedAuthority> authorities;
     private Map<String, Object> attributes;
 
-    public CustomOAuth2User(int userId, Provider userProvider, String userProviderId, String userName, Role role, Collection<? extends GrantedAuthority> authorities,Map<String, Object> attributes) {
+    public CustomOAuth2User(Long userId, Provider userProvider, String userProviderId, String userName, Role role, Collection<? extends GrantedAuthority> authorities,Map<String, Object> attributes) {
         this.userId = userId;
         this.userProvider = userProvider;
         this.userProviderId = userProviderId;
